backend/requirements.txt
setuptools==65.5.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.8.0,<3.0.0
python-dotenv==1.0.0
requests==2.31.0
huggingface-hub==0.19.4
numpy==1.26.2
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-docx==1.0.1
charset-normalizer==3.3.2
aiofiles==23.2.1
python-jose==3.3.0
fastapi-cors==0.0.6